+++ client/production.env
@@ -1,0 +1,1 @@
+API_URL=http://eventhunt.co
